6 v2 Done

+++ src/main/java/com/bookstore/business/bll/catalogmngmt/BookManagerServiceBean.java
@@ -59,7 +59,9 @@ public class BookManagerServiceBean {
  * @param book le livre a supprimé. Si book est null, l'opération de suppression n'est pas exécutée
  */
     public void deleteBook(Book book) { 
-        em.remove(book);
+        if(book.getId() != null){
+            em.remove(book);
+        }
     }
 
     /**
